Refactor: Created cpk folder and moved hooks inside
