+++ app/src/hooks/Cpk/useCpkAllowance.tsx
@@ -2,10 +2,10 @@ import { Signer, constants } from 'ethers'
 import { BigNumber } from 'ethers/utils'
 import { useCallback, useEffect, useState } from 'react'
 
-import { useConnectedWeb3Context } from '../contexts'
-import { ERC20Service } from '../services'
-import { pseudoNativeAssetAddress } from '../util/networks'
-import { RemoteData } from '../util/remote_data'
+import { useConnectedWeb3Context } from '../../contexts'
+import { ERC20Service } from '../../services'
+import { pseudoNativeAssetAddress } from '../../util/networks'
+import { RemoteData } from '../../util/remote_data'
 
 /**
  * Return the allowance of the given `signer` for the cpk.
