+++ app/src/hooks/Cpk/useCpkProxy.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react'
 
-import { useConnectedWeb3Context } from '../contexts'
-import { RemoteData } from '../util/remote_data'
+import { useConnectedWeb3Context } from '../../contexts'
+import { RemoteData } from '../../util/remote_data'
 
 /**
  * Return details about a user's proxy contract
