+++ app/src/hooks/index.ts
@@ -7,8 +7,8 @@ export { useCollateral } from './useCollateral'
 export { useCollateralBalance } from './useCollateralBalance'
 export { useCompoundService } from './useCompoundService'
 export { useContracts } from './useContracts'
-export { useCpkAllowance } from './useCpkAllowance'
-export { useCpkProxy } from './useCpkProxy'
+export { useCpkAllowance } from './Cpk/useCpkAllowance'
+export { useCpkProxy } from './Cpk/useCpkProxy'
 export { useDetectAdblocker } from './useDetectAdblocker'
 export { useDocumentDescription } from './useDocumentDescription'
 export { useDocumentTitle } from './useDocumentTitle'
